requirements.txt
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0
